// src/lib/neon/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
// ...
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          account_type:
            Database['public']['Enums']['profile_account_type'] | null;
          avatar_url: string | null;
          content_description:
            Database['public']['Enums']['content_description'] | null;
          content_display:
            Database['public']['Enums']['content_display'] | null;
          id: string;
          providers: string[];
          sources: Database['public']['Enums']['source'][] | null;
          username: string | null;
          username_history: Json | null;
        };
        Insert: {
          account_type?:
            Database['public']['Enums']['profile_account_type'] | null;
          avatar_url?: string | null;
          content_description?:
            Database['public']['Enums']['content_description'] | null;
          content_display?:
            Database['public']['Enums']['content_display'] | null;
          id: string;
          providers?: string[];
          sources?: Database['public']['Enums']['source'][] | null;
          username?: string | null;
          username_history?: Json | null;
        };
        Update: {
          account_type?:
            Database['public']['Enums']['profile_account_type'] | null;
          avatar_url?: string | null;
          content_description?:
            Database['public']['Enums']['content_description'] | null;
          content_display?:
            Database['public']['Enums']['content_display'] | null;
          id?: string;
          providers?: string[];
          sources?: Database['public']['Enums']['source'][] | null;
          username?: string | null;
          username_history?: Json | null;
        };
        Relationships: [];
      };
// ...
    };
// ...
    Enums: {
      content_description: 'FULL' | 'BRIEF' | 'NONE';
      content_display: 'TABLE' | 'TILES';
      image_processing_status:
        'pending' | 'processing' | 'completed' | 'failed';
      notification_type: 'system';
      playlist_sort_order: 'ascending' | 'descending';
      playlist_sorted_by: 'title' | 'datePublished' | 'playlistOrder';
      playlist_type: 'Public' | 'Private';
      profile_account_type: 'default' | 'admin';
      source:
        'giantbomb' | 'nextlander' | 'jeffgerstmann' | 'remap' | 'minnmax';
    };
    CompositeTypes: {
      username_history_entry: {
        username: string | null;
        used_from: string | null;
        used_until: string | null;
      };
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  'public'
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    keyof DefaultSchema['Enums'] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;